refactor(task-9): merge async & sync version of udp thermometer
