+++ smarthome/crates/libs/udp-thermometer/Cargo.toml
@@ -8,5 +8,12 @@ version           = "0.1.0"
 
 [dependencies]
 smarthome = { path = "../../libs/smarthome" }
+tokio = { version = "1.36.0", features = [
+    "io-util",
+    "macros",
+    "net",
+    "rt-multi-thread",
+    "time",
+] }
 [lints]
 workspace = true
