+++ smarthome/crates/libs/udp-thermometer/examples/basic_usage_async.rs
@@ -1,13 +1,19 @@
 use std::time::Duration;
+use tokio::time::sleep;
 
 use smarthome::devices::Gauge;
-use tokio::time::sleep;
-use udp_thermpmeter_async::UdpThermo;
+use udp_thermometer::{impl_async::AsyncVer, UdpThermo};
 
 #[tokio::main]
 async fn main() {
     let receiver_address = "127.0.0.1:4342";
-    let thermo = UdpThermo::new(receiver_address).await.unwrap();
+    let thermo = match UdpThermo::<AsyncVer>::new(receiver_address).await {
+        Ok(thermo) => thermo,
+        Err(e) => {
+            eprintln!("Failed to create UdpThermo: {e}");
+            return;
+        }
+    };
     for i in 0..120 {
         sleep(Duration::from_secs(1)).await;
         print!("Measure #{i} - ");
