+++ smarthome/crates/libs/udp-thermometer/examples/thermo_imitator_async.rs
@@ -1,12 +1,10 @@
-use std::{
-    net::{SocketAddr, UdpSocket},
-    thread,
-    time::Duration,
-};
+use std::{net::SocketAddr, time::Duration};
+use tokio::{net::UdpSocket, time::sleep};
 
 use udp_thermometer::temperature::TemperatureGenerator;
 
-fn main() {
+#[tokio::main]
+async fn main() {
     let receiver = std::env::args()
         .nth(1)
         .unwrap_or_else(|| "127.0.0.1:4342".into())
@@ -14,14 +12,14 @@ fn main() {
         .expect("valid socket address expected");
 
     let bind_addr = "127.0.0.1:4320";
-    let socket = UdpSocket::bind(bind_addr).expect("can't bind socket");
+    let socket = UdpSocket::bind(bind_addr).await.expect("can't bind socket");
 
     println!("Starting send temperature from {bind_addr} to {receiver}");
     for temperature in TemperatureGenerator::default() {
         let bytes = temperature.to_be_bytes();
-        if let Err(err) = socket.send_to(&bytes, receiver) {
+        if let Err(err) = socket.send_to(&bytes, receiver).await {
             eprintln!("can't send temperature: {err}");
         }
-        thread::sleep(Duration::from_secs(1));
+        sleep(Duration::from_secs(1)).await;
     }
 }
