+++ smarthome/crates/libs/udp-thermometer/src/impl_async.rs
@@ -1,46 +1,15 @@
-use std::{
-    fmt::{self, Display, Formatter},
-    sync::{
-        atomic::{AtomicBool, Ordering},
-        Arc, Mutex,
-    },
+use std::sync::{
+    atomic::{AtomicBool, Ordering},
+    Arc, Mutex,
 };
 
-use smarthome::devices::Gauge;
 use tokio::net::{ToSocketAddrs, UdpSocket};
 
-pub type Result<T> = core::result::Result<T, Error>;
+use crate::Result;
+use crate::{temperature::Temperature, UdpThermo};
+pub struct AsyncVer;
 
-pub type Error = Box<dyn std::error::Error>; // For early dev.
-
-#[derive(Default, Debug, Clone, Copy, PartialEq, PartialOrd)]
-pub struct Temperature(f32);
-
-impl From<f32> for Temperature {
-    fn from(value: f32) -> Self {
-        Self(value)
-    }
-}
-
-impl Display for Temperature {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl Temperature {
-    #[must_use]
-    pub fn to_be_bytes(self) -> [u8; 4] {
-        self.0.to_be_bytes()
-    }
-}
-
-pub struct UdpThermo {
-    temperature: Arc<Mutex<Option<Temperature>>>,
-    finished: Arc<AtomicBool>,
-}
-
-impl UdpThermo {
+impl UdpThermo<AsyncVer> {
     /// # Panics
     /// may panic
     /// # Errors
@@ -62,7 +31,7 @@ impl UdpThermo {
                 let mut buf = [0; 4];
                 let res = match socket.recv_from(&mut buf).await {
                     Ok((bytes_read, _)) if bytes_read == buf.len() => {
-                        Some(Temperature(f32::from_be_bytes(buf)))
+                        Some(Temperature::from_be_bytes(buf))
                     }
                     Ok((bytes_read, src_addr)) => {
                         eprintln!("only {bytes_read} bytes read from {src_addr}");
@@ -73,26 +42,17 @@ impl UdpThermo {
                         None
                     }
                 };
-                *temperature_clone.lock().unwrap() = res;
+                match temperature_clone.lock() {
+                    Ok(mut temperature) => *temperature = res,
+                    Err(e) => eprintln!("Failed to acquire lock: {e}"),
+                }
             }
         });
 
         Ok(Self {
             temperature,
             finished,
+            _sync: std::marker::PhantomData,
         })
-    }
-}
-
-impl Gauge<Temperature> for UdpThermo {
-    type R = Option<Temperature>;
-    fn get_measure(&self) -> Self::R {
-        *self.temperature.lock().unwrap()
-    }
-}
-
-impl Drop for UdpThermo {
-    fn drop(&mut self) {
-        self.finished.store(true, Ordering::SeqCst);
     }
 }
