--- smarthome/crates/libs/udp-thermpmeter-async/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-edition           = "2021"
-license.workspace = true
-name              = "udp-thermpmeter-async"
-version           = "0.1.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-smarthome = { path = "../../libs/smarthome" }
-tokio = { version = "1.36.0", features = [
-    "io-util",
-    "macros",
-    "net",
-    "rt-multi-thread",
-    "time",
-] }
-
-[lints]
-workspace = true
